+++ src/main/java/br/com/ecore/tom/exceptions/APIExceptionUtils.java
@@ -6,13 +6,13 @@ public class APIExceptionUtils {
 
   private APIExceptionUtils() {}
 
-  public static void assertValidPost(Long id) {
+  public static void assertValidPost(Integer id) {
     if (!Objects.isNull(id)) {
       throw new APIException("O id deve ser nulo para criação de novas entidades", 400L);
     }
   }
 
-  public static void assertValidPut(Long id) {
+  public static void assertValidPut(Integer id) {
     if (Objects.isNull(id)) {
       throw new APIException("Não é possível atualizar entidades sem id", 400L);
     }
